antelope/requirements.txt
absl-py==1.4.0
click==8.1.3
jaxlib==0.4.14
jax==0.4.14
numba==0.57.1
numpy==1.24.4
matplotlib==3.7.1
pandas==2.0.1
partikel==0.1.0
pytest==7.4.0
scikit-learn==1.2.2
seaborn==0.12.2
tensorflow-probability==0.19.0
torch==2.0.0
tqdm==4.65.0
